[Sdf, Usd, Hd, UsdImaging] First pass at fast update API with usda support and unit testing
This change introduces the concept of field handles to facilitate fast updates to attribute defaults and time samples.

Through investigation, @ashulzhenko and @aluk found two key bottlenecks when authoring attribute values to USD:

Digging through the datastore (SdfAbstractData subclass) to get to the location where we want to write.
Change processing which involves many copies of containers in order to robustly recompose and propagate edits.
Field handles provide authoring clients the ability to cache the datastore location at which to write. For now, we also use field handles to trigger a fastpath for change processing, which tracks only the property paths that have changed. This avoids overhead in constructing and copying map entries.

The tradeoff is that fast updates yield a subset of change notification that the normal authoring codepath provides. Currently, no clients depend on the full change notification, and only need to know the affected property paths.

Future work:

- Add usdc support.
- Thread field handle creation through the UsdStage, so that we can detect fields which have no dependent paths. Then we can restore the fastest codepath in Usd change processing which performs no copying of containers.
- Support querying time samples via field handles.
- Unit testing for authoring on scenegraph instancing with field handles.
- Allow authoring clients to opt in to each optimization of fast updates separately, i.e., taking advantage of field handles without opting into faster change processing, and vice versa.